+++ packages/logger/src/types.ts
@@ -1,0 +1,2 @@
+export * from "./EventsTypes";
+export * from "./LoggerTypes";